requirements.txt
Flask
torch
torchvision
Pillow
gdown
gunicorn
